refactored cell quality and archive
