+++ pyansys/archive.py
@@ -1,68 +1,100 @@
-"""
-Module to read ANSYS ASCII block formatted CDB files
+"""Module to read ANSYS ASCII block formatted CDB files
 """
 import sys
 import logging
+from functools import wraps
 
 import numpy as np
 from vtk import (VTK_TETRA, VTK_QUADRATIC_TETRA, VTK_PYRAMID,
                  VTK_QUADRATIC_PYRAMID, VTK_WEDGE, VTK_QUADRATIC_WEDGE,
                  VTK_HEXAHEDRON, VTK_QUADRATIC_HEXAHEDRON)
 import pyvista as pv
+import vtk
 
-from pyansys.vtk_helper import raw_to_grid
-from pyansys import _reader
+from pyansys._cellqual import cell_quality_float, cell_quality
+from pyansys.elements import valid_types
+from pyansys import _relaxmidside, _parser, _reader
 
+VTK9 = vtk.vtkVersion().GetVTKMajorVersion() >= 9
 
 log = logging.getLogger(__name__)
 log.setLevel('CRITICAL')
 
 
-class Archive(object):
+class Archive():
     """Read a blocked ANSYS archive file.
+
+    Reads a blocked CDB file and optionally parses it to a vtk grid.
 
     Parameters
     ----------
     filename : string
         Filename of block formatted cdb file
 
+    parse_vtk : bool, optional
+        When ``True``, parse the raw data into to VTK format.
+
+    force_linear : bool, optional
+        This parser creates quadratic elements if available.  Set
+        this to True to always create linear elements.  Defaults
+        to False.
+
+    allowable_types : list, optional
+        Allowable element types.  Defaults to all valid element
+        types in ``from pyansys.elements.valid_types``
+
+        See help(pyansys.elements) for available element types.
+
+    null_unallowed : bool, optional
+        Elements types not matching element types will be stored
+        as empty (null) elements.  Useful for debug or tracking
+        element numbers.  Default False.
+
+    Examples
+    --------
+    >>> import pyansys
+    >>> hex_beam = pyansys.Archive(pyansys.examples.hexarchivefile)
+    >>> hex_beam
     """
 
-    def __init__(self, filename, read_parameters=False, verbose=False):
+    def __init__(self, filename, read_parameters=False,
+                 parse_vtk=True, force_linear=False,
+                 allowable_types=None, null_unallowed=False,
+                 verbose=False):
         """ Initializes a cdb object """
         self.raw = _reader.read(filename, read_parameters=read_parameters,
                                 debug=verbose)
-        self.grid = None
 
-    def parse_vtk(self, force_linear=False, allowable_types=None,
-                  null_unallowed=False):
-        """Parses raw data into to VTK format.
+        self._grid = None
+        if parse_vtk:
+            self._grid = raw_to_grid(self.raw, allowable_types, force_linear,
+                                     null_unallowed)       
 
-        Parameters
-        ----------
-        force_linear : bool, optional
-            This parser creates quadratic elements if available.  Set
-            this to True to always create linear elements.  Defaults
-            to False.
+    @property
+    def grid(self):
+        if self._grid is None:
+            raise AttributeError('Archive must be parsed as a vtk grid.\n'
+                                 'Set `parse_vtk=True`')
+        return self._grid
 
-        allowable_types : list, optional
-            Allowable element types.  Defaults to all valid element
-            types in ``from pyansys.elements.valid_types``
+    @property
+    def quality(self):
+        cells = self.grid.cells
+        if cells.dtype != np.int64:
+            cells = cells.astype(np.int64)
 
-            See help(pyansys.elements) for available element types.
+        offset = self.grid.offset
+        celltypes = self.grid.celltypes
+        points = self.grid.points
 
-        null_unallowed : bool, optional
-            Elements types not matching element types will be stored
-            as empty (null) elements.  Useful for debug or tracking
-            element numbers.  Default False.
+        if points.dtype == np.float64:
+            return cell_quality(cells, offset, celltypes, points)
+        return cell_quality_float(cells, offset, celltypes, points)
 
-        Returns
-        -------
-        grid : vtk.vtkUnstructuredGrid
-            VTK unstructured grid from archive file.
-        """
-        self.grid = raw_to_grid(self.raw, allowable_types, force_linear, null_unallowed)
-        return self.grid
+    @wraps(pv.plot)
+    def plot(self, *args, **kwargs):
+        """Plot the ANSYS grid"""
+        self.grid.plot(*args, **kwargs)
 
 
 def chunks(l, n):
@@ -572,3 +604,154 @@ def write_cmblock(filename, items, comp_name, comp_type, digit_width=10):
         open(filename, 'w').write(text)
     else:
         filename.write(text)
+
+
+def raw_to_grid(raw, allowable_types, force_linear, null_unallowed):
+    """Parses raw data into to VTK format.
+
+    Parameters
+    ----------
+    force_linear : bool, optional
+        This parser creates quadratic elements if available.  Set
+        this to True to always create linear elements.  Defaults
+        to False.
+
+    allowable_types : list, optional
+        Allowable element types.  Defaults to all valid element
+        types in ``from pyansys.elements.valid_types``
+
+        See help(pyansys.elements) for available element types.
+
+    null_unallowed : bool, optional
+        Elements types not matching element types will be stored
+        as empty (null) elements.  Useful for debug or tracking
+        element numbers.  Default False.
+
+    Returns
+    -------
+    grid : vtk.vtkUnstructuredGrid
+        VTK unstructured grid from archive file.
+    """
+    # Convert to vtk style arrays
+    if allowable_types is None:
+        allowable_types = valid_types
+    else:
+        assert isinstance(allowable_types, list), \
+               'allowable_types must be a list'
+        for eletype in allowable_types:
+            if str(eletype) not in valid_types:
+                raise Exception('Element type "%s" ' % eletype +
+                                'cannot be parsed in pyansys')
+
+    # construct keyoption array
+    keyopts = np.zeros((10000, 20), np.int16)
+
+    for keyopt_key in raw['keyopt']:
+        for index, value in raw['keyopt'][keyopt_key]:
+            keyopts[keyopt_key, index] = value
+
+    # parse raw output
+    parsed = _parser.parse(raw, force_linear, allowable_types,
+                           null_unallowed, keyopts)
+
+    cells = parsed['cells']
+    offset = parsed['offset']
+    cell_type = parsed['cell_type']
+    numref = parsed['numref']
+    enum = parsed['enum']
+
+    # Check for missing midside nodes
+    if force_linear or np.all(cells != -1):
+        nodes = raw['nodes'][:, :3].copy()
+        nnum = raw['nnum']
+        angles = raw['nodes'][:, 3:]
+    else:
+        mask = cells == -1
+
+        nextra = mask.sum()
+        maxnum = numref.max() + 1
+        cells[mask] = np.arange(maxnum, maxnum + nextra)
+
+        nnodes = raw['nodes'].shape[0]
+        nodes = np.zeros((nnodes + nextra, 3))
+        nodes[:nnodes] = raw['nodes'][:, :3]
+
+        # Set new midside nodes directly between their edge nodes
+        temp_nodes = nodes.copy()
+        _relaxmidside.reset_midside(cells, cell_type, offset, temp_nodes)
+        nodes[nnodes:] = temp_nodes[nnodes:]
+
+        # merge nodes
+        new_nodes = temp_nodes[nnodes:]
+        unique_nodes, idxA, idxB = unique_rows(new_nodes)
+
+        # rewrite node numbers
+        cells[mask] = idxB + maxnum
+        nextra = idxA.shape[0]
+        nodes = np.empty((nnodes + nextra, 3))
+        nodes[:nnodes] = raw['nodes'][:, :3]
+        nodes[nnodes:] = unique_nodes
+
+        angles = np.empty((nnodes + nextra, 3))
+        angles[:nnodes] = raw['nodes'][:, 3:]
+        angles[nnodes:] = 0
+
+        # Add extra node numbers
+        nnum = np.hstack((raw['nnum'], np.ones(nextra, np.int32) * -1))
+
+    # Create unstructured grid
+    if VTK9:
+        grid = pv.UnstructuredGrid(cells, cell_type, nodes)
+    else:
+        grid = pv.UnstructuredGrid(offset, cells, cell_type, nodes)
+
+    # Store original ANSYS element and cell information
+    grid.point_arrays['ansys_node_num'] = nnum
+    grid.cell_arrays['ansys_elem_num'] = enum
+    grid.cell_arrays['ansys_elem_type_num'] = parsed['etype']
+    grid.cell_arrays['ansys_real_constant'] = parsed['rcon']
+    grid.cell_arrays['ansys_material_type'] = parsed['mtype']
+    grid.cell_arrays['ansys_etype'] = parsed['ansys_etype']
+
+    # Add element components to unstructured grid
+    for comp in raw['elem_comps']:
+        mask = np.in1d(enum, raw['elem_comps'][comp],
+                       assume_unique=True)
+        grid.cell_arrays[comp.strip()] = mask
+
+    # Add node components to unstructured grid
+    for comp in raw['node_comps']:
+        mask = np.in1d(nnum, raw['node_comps'][comp],
+                       assume_unique=True)
+        grid.point_arrays[comp.strip()] = mask
+
+    # Add tracker for original node numbering
+    ind = np.arange(grid.number_of_points)
+    grid.point_arrays['origid'] = ind
+    grid.point_arrays['VTKorigID'] = ind
+
+    # store node angles
+    grid.point_arrays['angles'] = angles
+    return grid
+
+
+def check_raw(raw):
+    """ Check if raw data can be converted into an unstructured grid """
+    try:
+        raw['elem'][0, 0]
+        raw['enum'][0]
+    except Exception:
+        # return True
+        raise Exception('Invalid file or missing key data.  ' +
+                        'Cannot parse into unstructured grid')
+
+
+def unique_rows(a):
+    """ Returns unique rows of a and indices of those rows """
+    if not a.flags.c_contiguous:
+        a = np.ascontiguousarray(a)
+
+    b = a.view(np.dtype((np.void, a.dtype.itemsize * a.shape[1])))
+    _, idx, idx2 = np.unique(b, True, True)
+
+    return a[idx], idx, idx2
